Moved golang code into project root directory
