+++ cache.go
@@ -1,7 +1,7 @@
 package gubernator
 
 import (
-	"github.com/mailgun/gubernator/golang/cache"
+	"github.com/mailgun/gubernator/cache"
 	"github.com/mailgun/holster"
 )
