+++ cmd/gubernator/main.go
@@ -16,8 +16,8 @@ import (
 	etcd "github.com/coreos/etcd/clientv3"
 	"github.com/ghodss/yaml"
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
-	"github.com/mailgun/gubernator/golang"
-	"github.com/mailgun/gubernator/golang/cache"
+	"github.com/mailgun/gubernator"
+	"github.com/mailgun/gubernator/cache"
 	"github.com/mailgun/holster"
 	"github.com/mailgun/holster/clock"
 	"github.com/mailgun/holster/etcdutil"
