+++ functional_test.go
@@ -7,8 +7,8 @@ import (
 	"testing"
 	"time"
 
-	guber "github.com/mailgun/gubernator/golang"
-	"github.com/mailgun/gubernator/golang/cluster"
+	guber "github.com/mailgun/gubernator"
+	"github.com/mailgun/gubernator/cluster"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
