+++ hash.go
