+++ python/setup.py
@@ -25,8 +25,8 @@ setup(
     author="Derrick J. Wippler",
     author_email='thrawn01@gmail.com',
     url='https://github.com/mailgun/gubernator',
-    package_dir={'': 'python'},
-    packages=find_packages('python', exclude=['tests']),
+    package_dir={'': '.'},
+    packages=find_packages('.', exclude=['tests']),
     install_requires=requirements,
     license="Apache Software License 2.0",
     python_requires='>=2.7',
