--- stub.go
@@ -1,4 +1,0 @@
-package gubernator
-
-// This is a dummy package for go mod to work properly
-var Stub string
